solana_remote_wallet

Module ledger

Structs

LedgerSettings

LedgerWallet

Ledger Wallet device

Enums

PubkeyDisplayMode

Functions

get_ledger_from_info

Choose a Ledger wallet based on matching info fields

is_valid_ledger

Check if the detected device is a valid Ledger device by checking both the product ID and the vendor ID