solana_remote_wallet

Module remote_wallet

Structs

Device

RemoteWallet device

RemoteWalletInfo

Remote wallet information.

RemoteWalletManager

Collection of connected RemoteWallets

Enums

RemoteWalletError

Remote wallet error.

RemoteWalletType

Remote wallet convenience enum to hold various wallet types

Traits

RemoteWallet

RemoteWallet trait

Functions

initialize_wallet_manager

is_valid_hid_device

Helper to determine if a device is a valid HID

maybe_wallet_manager