solana_remote_wallet::remote_wallet

Trait RemoteWallet

pub trait RemoteWallet<T> {
    // Provided methods
    fn name(&self) -> &str { ... }
    fn read_device(
        &mut self,
        dev_info: &T,
    ) -> Result<RemoteWalletInfo, RemoteWalletError> { ... }
    fn get_pubkey(
        &self,
        derivation_path: &DerivationPath,
        confirm_key: bool,
    ) -> Result<Pubkey, RemoteWalletError> { ... }
    fn sign_message(
        &self,
        derivation_path: &DerivationPath,
        data: &[u8],
    ) -> Result<Signature, RemoteWalletError> { ... }
    fn sign_offchain_message(
        &self,
        derivation_path: &DerivationPath,
        message: &[u8],
    ) -> Result<Signature, RemoteWalletError> { ... }
}

RemoteWallet trait

Provided Methods

fn name(&self) -> &str

fn read_device( &mut self, dev_info: &T, ) -> Result<RemoteWalletInfo, RemoteWalletError>

Parse device info and get device base pubkey

fn get_pubkey( &self, derivation_path: &DerivationPath, confirm_key: bool, ) -> Result<Pubkey, RemoteWalletError>

Get solana pubkey from a RemoteWallet

fn sign_message( &self, derivation_path: &DerivationPath, data: &[u8], ) -> Result<Signature, RemoteWalletError>

Sign transaction data with wallet managing pubkey at derivation path m/44'/501'/<account>'/<change>'.

fn sign_offchain_message( &self, derivation_path: &DerivationPath, message: &[u8], ) -> Result<Signature, RemoteWalletError>

Sign off-chain message with wallet managing pubkey at derivation path m/44'/501'/<account>'/<change>'.